fix: copy file 삭제 및 최근 소식 type 제대로 지정
