--- frontend/src/query/get/useGetLatestNews.tsx
@@ -1,36 +1,0 @@
-import { QUERY_KEYS } from '@constants/keys';
-import { useSuspenseQuery } from '@tanstack/react-query';
-import request from '@utils/request';
-
-interface News {
-  content: [
-    {
-      articleId: number;
-      grade: string;
-      name: string;
-      date: string;
-      imageExistence: boolean;
-      subject: string;
-    },
-  ];
-}
-
-function useGetLatestNews() {
-  const fetchGetLatestNews = async () => {
-    const response = await request<null, News, null>({
-      uri: '/api/main/latest-news?page=0&size=3',
-      method: 'get',
-    });
-
-    return response.data;
-  };
-
-  const { data } = useSuspenseQuery({
-    queryKey: [QUERY_KEYS.GET_LATEST_NEWS],
-    queryFn: fetchGetLatestNews,
-  });
-
-  return { data };
-}
-
-export default useGetLatestNews;
